backend/requirements.txt
Flask==3.0.3
flask-cors==5.0.0
langchain==0.3.2
langchain-openai==0.2.2
langchain-google-genai==2.0.0
langchain-community==0.3.1
python-dotenv==1.0.1